rpk topic consume: redo for franz-go
This rewrites the consume code to be specific to franz-go. As with the
prior commit, the new code does effectively no work other than parsing
options and beginning a consume loop. Since we have integration tests
for consuming, and because franz-go has no mock client, the tests
ensuring a produce actually happens are removed.

The default output format is still JSON, but there is one slight change
to the output, described below. Other than that, the output formatting
options are now quite overpowered.

Changes:

- commit has been deprecated; now, whenever consuming as a group,
committing is enabled by default. There isn't much reason to not commit
as a group, and there will be even less reason when we implement the
DeleteGroups API.

- the default format's json has had its "size" field renamed to
"value_size". This is clearer as to what's being suppressed when using
--meta-only.

- the --offset flag now supports relative start & end, as well as
ranges.