rptest rpk: minor changes for prior commits, improvement
- produce: use old format %v

The new format is %v\n, which is incompatible with the old format. To
avoid changing any surrounding tests, we can just use the old format
directly.

- produce: avoid adding dedicated --brokers flag

Produce calls into _run_topic which also adds the brokers flag, so we
can remove it from the produce function.

- produce: explicitly disable compression

To get exact parity with the old behavior.

- produce: add --delivery-timeout 4.5s

The old rpk used the following combination of settings:
  * 1s produce timeout
  * 2s backoff
  * 3 retries max
which would result in 9s total to try producing. The rpk produce test
code uses a 5s timeout. We will use a 4.5s delivery timeout to squeak in
under the 5s limit.

- execute: use subprocess.communicate

From online reading, using p.wait can deadlock if the subprocess writes
before we are reading, which will manifest in timeout errors.
p.communicate is basically much better all around, and we can avoid
writing our our output to an intermediate file and avoid polling
completion, so we may as well just switch over, cut code a good deal,
and avoid the potential to deadlock.

- execute: capture stderr in pipe in output logging as well

stderr was previously printed directly to logs. We may as well capture
stderr and include it in our error message, so that the logs will be a
part of the build artifacts (which is better overall).