util/named_type: Create rvalue overload to prevent dangling ref
named_type is designed to behave as a built-in value type.

The 1st specialisation for arithmetic types returns by copy.
The 2nd specialisation returns the underlying value by const&

The 2nd version is often based on strings, so this makes sense
as an optmisation to reduce allocations due to copying..

Prior to this change, it's possible to create a dangling reference:
```
using t = named_type<ss::sstring, struct int_t_alias_test_module>;
static constexpr std::string_view str{"This shouldn't dangle"};
auto f = []() { return t{ss::sstring{str}}; };
const auto& r = f()();
```

The `const auto& r` doesn't perform lifetime extension of the
underlying value of the  named_type, as might be expected.

To satisfy the expectation, introduce ref-qualified accessors:
if the named_type is an rvalue, move the value out (by value).
if the named_type is an lvalue, return value by const ref.

This provides the expected behaviour and more closely mimics the
behaviour of built-in value types.