k/group_manager: keep group pointer alive while doing group actions
Many of `kafka::group` operations are returning futures. When group is
deleted it is kept alive by the shared pointer in a call site. When
group operation is asynchronous and returns future we must keep the
pointer alive until operation is finished. This way we will prevent
state from being deleted.

Signed-off-by: Michal Maslanka <michal@vectorized.io>