rpk topic consume: obey -n, commit on shutdown
In the franz_produce_consume PR, I forgot to take account of -n, meaning
we would not quit after a user specified limit. We now do, but to
support this we now deeply loop on metches manually so so that we can
early return deep inside the looping.

As well, this fixes one long standing flaw of kcl: by relying on
autocommitting, we would finish consuming so quickly that we would not
commit before leaving. Now, we mark every record we want to commit each
loop, and more importantly, before returning from our consume loop.
Autocommitting uses these marks to commit (instead of the default
autocommit behavior, allow-everything-previously-polled-when-polling).

Leaving the group properly on shutdown ensures that our commit works.

Now, running rpk a few times, we can see
- offset starts at 123 for a new group
- we stop after one consume
- each repeated invocation uses the commit from the prior run,
regardless of input offset:

$ rpk topic consume foo -o 123 -n 1 -g group -f '%o\n'
123
$ rpk topic consume foo -o 123 -n 1 -g group -f '%o\n'
124
$ rpk topic consume foo -o 123 -n 1 -g group -f '%o\n'
125
$ rpk topic consume foo -o 123 -n 1 -g group -f '%o\n'
126