plugin: prefix plugins with '.'
When a plugin is installed, it is moved into a user's path. If plugins
have the prefix "rpk-", then typing `rpk` and hitting tab will not show
commands for rpk. Instead, it'll show that there are two potential
commands to run: rpk, or rpk-<plugin>.

By prefixing with a dot, we tab directly into rpk.

The main downside with this is that now if a user types `.` and hits
tab, the plugin will show. But, this is some small thing.

Potentially, we can install plugins in a dedicated path in the future,
and have rpk look there first. The path could be overridden with --dir.