kafka: prevent early group ops before partition recovery
The group manager maintains an index of attached_partitions (partition
pointer plus state) for each partition in the consumer group topic that
maps to the current shard. There are two pieces of state that important
in the context of this commit:

  - attached_partition.loading
  - current partition leadership

validate_group_state
--------------------

An attached_partition object with `loading=false` is created when the
partition manager notifies the group manager that a consumer group
partition is now managed on the current core.

Once this attached_partition is registered, then operations may observe
it in `validate_group_status` which guards group operations. one thing
it does is returns `not_coordinator` if the partition a group maps to is
not the leader OR the partition is in the loading state.

Because a freshly created attached_partition object has loading=false,
if the partition happens to be the leader then validate_group_status
will permit the operation. When this occurs _prior_ to the consumer
group log being replayed a group operation will not observe up-to-date
state.

Therefore this commit switches the initial value of loading to be true
so that ops that arrive early are turned away.

list_groups
-----------

However, this poses a problem for list_groups because it will return a
retriable error if any attached partition is in the loading state. since
we don't detach partitions when we lose leadership, then we end up in a
situation where list_groups always returns a retriable error because an
attached partition on a core is not a current leader.

Thus we want loading to represent a state that eventually turns into
non-loading (so we don't block list_groups). But we had been delivering
leadership change notifications prior to attaching partitions which
woudl cause the leadership notification to be dropped.

notifications
-------------

When the kafka group manager register for partition change notifications
(e.g. attach/detach) and leadership notifications any existing state is
delivered synchronously to the upcalls at registration time.

Prior to this change we registered leadership change first, which means
that if any partitions existed then leadership change would have been
dropped on the floor because it wouldn't be until we registered for
partition changes that we'd actually attach the partition.

If the partition had managed to already be running / healthy by the time
kafka group manager started during the redpanda start-up sequence then
it may miss a leaderhsip notification entirely.

final setup
-----------

- Newly attached partitions have loading=true. This means some early ops
  are correctly rejected until recover completes.

- List groups will return a retriable error after a partition is newly
  attached.

- Leadership notifications are delivered such that when we lose
  leadership loading=false. In this case list_groups proceeds, and
  validate_group_status relies on leadership check to reject operations
  (rather than loading).

- If we gain leadership, then loading=true is set (if we were in the
  previous state of loading=false, for example because we weren't the
  leader) and finally loading=false is set after recovery completes and
  any operations receiving retriable errors could proceed.

Signed-off-by: Noah Watkins <noah@vectorized.io>
(cherry picked from commit b54ff1dea7d2d9ea00a4df94d8570a14ab264d7a)