kafka: always succeed in loading consumer group partition
This case exists in Kafka which we also check for:

<kafka>
// if the cache already contains a group which should be removed, raise an error. Note that it
// is possible (however unlikely) for a consumer group to be removed, and then to be used only for
// offset storage (i.e. by "simple" consumers)
</kafka>

both redpanda and kafka throw an exception here. however, in kafka the
exception caught, logged, but the partition is finally cleared the loading
flag. in redpanda, the thrown exception prevented clearing the loading flag.
this could then lead to a stuck partition.

Signed-off-by: Noah Watkins <noah@vectorized.io>