config: add File to Config
The debug bundle command grabs the raw underlying redpanda.yaml file on
the system. Previously, there was some hackery to work with the viper
system to grab that file.

We can convert debug bundle to using ParamsFromConfig, and we can save
the decoded file in a field in the Config struct itself. This will
simplify things.