(split) switch the rest of sarama to franz-go
This switches all remaining references of sarama in code that isn't
scheduled for deletion to franz-go. This contains basically a few odds
and ends all over.

The main two code changes are in prometheus generation and container
starting, but those switches are tiny and easy. We just use the new kadm
client.

This requires a small kadm dep bump to use the new ListBrokers function.