coproc/tests: Use the client_fetch_batch_reader
- To remove code duplication and creating a manual consumption loop, the
logic in kafka::client::client_fetch_batch_reader can be used

- This involves changing the API of drain() it will also be renamed to
consume() to avoid confusion.

- Previously the test harness would read to the end of the log if an
empty read was detcted.

- Now a list offsets request is made before the actual call to fetch, so
it is not necessay to return an optional<batch_reader>, either a batch
reader is returned, or an exception is thrown if the underlying topic is
never created.

- Furthermore the API is modified to include 'last_offset' which
replaces the 'limit' parameter. This is because the
client_fetch_batch_reader reads up until a given offset, not a number of
records.