k8s:fix values being parsed as scientific notation
This boils down to how rpk config manager works. If you don't provide
single, it probably treats it as json and in the end viper when setting
the value, converts it to scientific notation (it does that for certain
numbers). Providing single as format resolved this.

(cherry picked from commit c512895fa161d2fcbedda281362a572b7afc7982)