raft: step down immediately on successful leader transfer
Once we have passed all our preparation and pre-checks,
and received an ack to the timeout_now rpc, we can be
reasonably sure that the new leader will proceed to take
up their role.

This is a tradeoff:
- Stepping down has risk that we that something goes wrong
  with new election and the partition stops servicing writes
  for a few seconds while priorities adjust a new leader is found.
- Not stepping down has risk that we will receive writes after
  timeout_now is complete, causing the old leader's log to advance,
  and the new leader to potentially lose its election.

On balance, it's better to have leadership transfer work reliably
(except when something really goes wrong) rather than having it
only work most of the time (even when the system is healthy).

Signed-off-by: John Spray <jcs@vectorized.io>