coproc: Add functionality to wait on idle script
- This patch attempts to solve the issue of knowing when to stop
consuming from a materialized topic.

- The problem is pushing to an input topic is an asyncronous action that
has no dependency on the event we want to wait on i.e. when the
materialized topic has all records written into it.

- Instead of busy waiting, now the ability to wait on when the coproc
fiber is in an idle state will be the event to use to know when to begin
consuming the test result data.