coproc/tests: Wait on scripts before consuming
- Previously drain() (what consume() was named before this patch series)
accepted a parameter called 'limit' and the drain poll loop would not
return until limit was reached or some timeout was exceeded.

- Now that the kafka client has been introduced of favor of this "poll
until limit reached" behavor, a new system must be introduced to make
consume() behave as it did previously, i.e. wait until the materialized
log has all records written to it.

- As an alternative to polling this commit works by waiting until
all active scripts hit the idle state. This will work becuase a script
can only be in the idle state if, it never recieved work, or if it
already processed all data on its respective input topics and is
'up to date' with the latest offset of all input topics.

- Therefore its safe to assume that when all scripts have reached their
respective idle state, the consume() method can continue to query
highest offset, and then make the call to fetch.