c/controller_backend: never skip creating partition when moved x-cores
When partition is moved cross cores there is a possibility that
partition will not get created on target core. This may happen when
`update_finished` operation has already been replicated for a given
update and partition wasn't yet shutdown on the core that it has to be
moved from.

Example:

We assume that partition was moved x-cores on each node in replica set
but not moved cross nodes.

1) update finishes on one of the nodes being member of replica set
2) `controller_backend` receives update finished delta
3) move target shard processes update_partition command and skips
   waiting for it to finish since operation is finished (update_finished
   operation is present)
4) partition is shutdown on the originating core

Fixed problem by not skipping cross core partition creation operation
(point 3) when waiting for x-core request from originating core. This
way 4 will be executed on originating core and partition will eventually
be created on target core.

Fixes: #2588

Signed-off-by: Michal Maslanka <michal@vectorized.io>