c/types: added version to incremental_topic_updates
Added version field to `cluster::incremental_topic_updates`. We use the
same mechanism as when introduced  in `cluster::topic_configuration`
serialization. New format version is negative and serialized as as first
field in resulting buffer. When deserializing type we peek the version
from the buffer if the value is negative we deal with new format
otherwise first field indicates optional presence and it can have only
two values i.e. either 0 or 1

Signed-off-by: Michal Maslanka <michal@vectorized.io>