cluster: shard_table add sanity checks
- Now that the shard_table is used to hold state about non replicable
topics, the class is permitted to have its two caches contain different
values when previously this was never true.

- To ensure updates/deletions using one API does not clash with another,
assertions are added as a sanity check.

- The assertion works by additing a boolean to the shard_revision
struct. This bool is set to true when a shard_revision is created via an
APIs without group_id . Upon the call to erase it is checked.