r/consensus: limiting last applied checkpoint to committed offset
In controller, to speed up bootstrapping we relay on the invariant that
the offset which was last applied to state machine is readable when
starting. Raft doesn't require that assumption to maintain correctness
it is enough for given message to be flushed on majority on nodes
(not necessary a leader). This may lead to the situation in which last
applied entry will not be readable after restart since it wasn't yet
flushed before it was applied. To be able to relay on the invariant that
last applied offset is readable we limit it to committed offset. This
way we will delay writing last applied offset until it is guaranteed to
be readable after restart.

Signed-off-by: Michal Maslanka <michal@vectorized.io>