admin api: cancel all but the successful request
net/http does not read the response body when returning a successful
response. If we cancel the context immediately after we receive the
response, we kill reading the body of the successful response.

Instead, we want to cancel all *other* requests that are destined to be
unsuccessful, and leave our own context fine and valid so we can read
the body.