cloud_storage: Reproduce and fix infinity cycle
Fetch can stuck in some cases (e.g. if estimated end of the one segment
is larger of the estimated end of the next).

This commit fixes the problem by using different search algorithm for
segments and by tracking redpanda offsets in record batch reader
differently.

It also adds bunch of unit-tests that reproduce this failure.