coproc: Transactional writes
- When multiple replies to write to different materialized logs share
the same source topic, if a retry is initiated between writes, the
offset will move forward for the source topic even through a retry is
initiated when the second errors.

- This means after retry one of the two destination topics will have more
records then expected.

- Solution is to group replies into transactions, where the key is the
input topic. This means the offset for this input topic will not be
moved until all operations under it succeed.

- As a side effect this change should also increase write throughput as
more work is done concurrently. Previously requests were processed
sequentially.

- Fixes #2771
- Fixes #2676
- Fixes #2613