cloud_storage: Add randomized test
Add randomized unit-test that generate random set of partitions with
configuration batches. The test fetches data the same way kafka layer
does the same thing. It creates a reader to fetch a chunk of data and
then it updates reader config and re-creates a reader to fetch the next
chunk. The process repeats until the partition is scanned.

The max_bytes value is fixed. This is fine because the data are
randomized (the individual batches are totally random, have different
number of records and sizes) so it's highly unlikely that the readers
stride will match a batch boundary.