acl delete: handle top level errors
Previously, we only printed the inner filter match error. The filter
itself can have an error.

This borrows logic that exists in `acl create`: we match the created
filter with the output result; if the result fails, we print the filter
that caused the failure.

This is better than previous, where the filter error would result in no
matches and then the user would not have any notification.