cluster: introduce archival_metadata_stm and integrate with partition
archival_metadata_stm allows storing archival manifest (a set of
segments archived to cloud storage) in the archived partition log
itself. This is needed to 1) avoid querying cloud storage on partition
startup and 2) to replicate metadata to raft followers so that they
can decide which segments can be safely evicted.