k8s/test: add `startKIND: true` to kuttl config
Remove the assumption that `kind create cluster` is executed before the
invocation to kuttl. This was originally added because `kind create` was
causing problems and it was necessary to inspect the output of the kind
command separately.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>
(cherry picked from commit 9ea4ef82f94d0595930c5306f676e337b204f5b8)