k8s: Fix complex types in additional configuration
With changes in golang dependencies one of bump broke viper and it's
ability to set the complex type.

This commit tries to guess tree basic types and use manager single parameter
in Set function. The problem is caused in the type set by viper library.
We expect this value to be int, but viper deduces this as float64. The float64
is interpreted by yaml package in scientific notation.