cloud_storage: Fix hydration in the remote_segment
Instead of hydrating on demand create a background fiber that hydrates
on demand. The fiber tracks all requesters and can notify them when the
hydration is done or when the error occured. This eliminates the problem
that causes subsequent read request to queue and restart the download
over and over again if we don't have enough bandwidth.

Also, disable the prefetch in remote_partition since it uses too much
resources.