r/vote_stm: reset leader_id in prevote already
Previously we reset leader_id in `vote_stm` when there is no majority of
the nodes reachable leader_id will never be reset since a node will
never be able to pass after the prevote phase. When node is sending
prevotes it is already an indication that the node didn't received the
heartbeats i.e. there is no active leader.

Signed-off-by: Michal Maslanka <michal@vectorized.io>