cloud_storage: Force reconnect on any error
Previously the code in the remote api pkg tried to be smart about error
handling by picking which errors should lead to reconnections and which
shuldn't. This led to situation when unrelated error was mistaken with
the transport error that shouldn't result in reconnection which caused a
bug.

This commit makes remote api reconnect after every failure since it's
not always possible to distinguish between transport errors and errors
generated by callback.

Also, verbocity of error handling is increased in some cases.