tests: update partition_movement_test for admin API changes
The partition move request handler used to return 400 for
in_progress partitions.  That has been changed to 503, to
better reflect that the request is well formed but the
server can't action it right now.

To receive that 503 in the test, it is also necessary to
disable the admin API client's new default retry logic.

Signed-off-by: John Spray <jcs@vectorized.io>