cloud_storage: Add remote_partition class
The remote_partition represetns a collection of remote_segment
instances that correspond to the uploaded segments of the partition.

The remote_segment instances are initialized lazily. On startup the
partition contains only a manifest and the collection of object keys.
When the segments are accessed by the fetch API the remote_segment
instances are materialized. The materialized instances are used to
create a model::record_batch_reader instances that access data from
S3. Every remote_partition instance has a timer which is used to run
cleanup periodically. The cleanup process offloads all materialized
segments that wasn't used long enough.

The remote_partition also implements record_batch_reader_impl which is
used to create a model::record_batch_reader that can scan multiple
remote segments.

This commit also contains fixes from shadow-indexing-merged:
- 6bb1a5b fix use-after-free
- 591b01e get rid of remote_partition::reader_state
- 61ad458 reproduce and fix infinite cycle
- 21a238f fix hydration in the remote segment
- 171ab0c remove timeout from work-item queue
- f916b8c prevent remote-segment from being evicted
- f00d5bf add availability check

Co-authored-by: Alexey Zatelepin <ztlpn@vectorized.io>