gha: change ccache key for redpanda workflow
Use git ref as part of the key and remove timestamp. The current limit
from Github is 5GB for all cached files across a repository. With this
change, there is only one ccache file uploaded per branch, instead of
one for each distinct build. This results in having one for the main
branch, as well as one per PR.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>