c/health_monitor: implemented cluster health monitor
Implemented cluster health monitor. The cluster health monitor consist
of two parts. The `cluster::health_monitor_backend` which is a central
part of the monitor. Health monitor backend is instantiated only on core
0. Health monitor backend can either collect cluster health periodically
(this happen on the raft0 leader) or ask leader health monitor for
cluster health report and cache it locally. This way state of cluster
health monitor is eventually consistent. Health monitor frontend is an
entry point for the health monitor subsystem. The frontend component
lives on every shard and expose convenient interface to interact with
health monitor.

Signed-off-by: Michal Maslanka <michal@vectorized.io>