coproc: Fix retry handling bugs
- Coproc failed to handle a number of critical cases with respect to
bumping the input topic offset.

- New logs appearing after processing started would always be processed
from the current offset instead of starting from 0 or latest/stored.

- If a failure occurred during processing of replies reply would occur
and double writes to a materialized log could happen.

- The solution is to maintain a cache of materialized ntps for each
coprocessor, denoting if the partition is up to date or not. On the read
side if at least one of these topics is not up to date, then replay. On
the write side, if a reply is detected and the partition is not behind,
then do not write.

- Fixes #2771
- Fixes #2613
- Fixes #2676