compatibility: modified PageViews and Wikipedia
Previously, the PageView example in ducktape only checked for a subset
of the regions the output. This resulted in an infrequent test failure
because the output had data outside of the subset. Fix by checking for
all possible regions.

The Wikipedia example also was checking for a subset of the data.
Instead check for all possible outputs.