k/metadata: use heuristic when there is no leader id available
We use simple heuristic to tolerate isolation of a node hosting both
partition leader and follower.

Kafka clients request metadata refresh in case they receive error that
is
related with stale metadata - f.e. NOT_LEADER. Metadata request can be
processed by any broker and there is no general rule for that which
broker to choose to refresh metadata from. (f.e. Java kafka client uses
the
broker with active least loaded connection.) This may lead to the
situation
in which client will ask for metadata always the same broker. When that
broker is isolated from rest of the cluster it will never update its
metadata
view. This way the client will always receive stale metadata.

This behavior may lead to a live lock in an event of network partition.
If
current partition leader is isolated from the cluster it will keep
answering
with its id in the leader_id field for that partition (according to
policy
where we return a former leader - there is no leader for that broker, it
is a
candidate). Client will retry produce or fetch request and receive
NOT_LEADER
error, this will force client to request metadata update, broker will
respond
with the same metadata and the whole cycle will loop indefinitely.

In order to break the loop and force client to make progress we use
following
heuristics:

1) when current leader is unknown, return former leader (Kafka behavior)
2) when current leader is unknown and previous leader is equal to
   current
   node id select random replica_id as a leader (indicate leader
isolation)

With those heuristics we will always force the client to communicate
with the
nodes that may not be partitioned.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 2a0269f142950acfc562f1cedc03b85c117edd7c)