cloud_storage: Fix segment hydration
Hydrate segment using the background fiber instead of doing this
on demand. This eliminates races since the hydration is done
sequentially in the background. Old approach was to start hydration
when the 'data_stream' method is called. The 'hydrate' method checked
the state of the object in the cache and if it's not there the method
started the download. If the download is in-progress the method is
pooling the state of the cache untill the state has changed. This
introduced the opportunity for the race conditions since arbitrary
time could pass between the check and the actual hydration. This
is because s3 connection pool can block if it's empty. So more than
one fiber can go into the 'hydrate' method and start hydration which
resulted in multiple downloads of the same object which in turn
exausted the network bandwidth.

Another possible problem is that the remote_segment could be evicted
by the remote_partition. But in this case the in-progress downloads
doesn't stop immediately but continue until completion or error. At the
same time the new version of the remote_segment for the same base offset
could be created and start another parallel download. This makes
synchronization tricky because it should take into account not one
remote_segment but all that have the same base offset because the cache
is a global state shared between all remote_segment instances.

The new approach is to perform hydration using the background fiber.
When the 'data_stream' method is called it adds a promise to the queue
and notifies the condition variable. The background fiber then wakes up,
checks the cache state and hydrates the segment if needed. At the end it
notifies the waiters. This approach works even if the caller was deleted
(which can happen if hydration takes long time).

The 'remote_segment' also has a way to check if the hydration is in
progress. This is used in the 'remote_partition' to prevent eviction of
such segments.

This commit also contains fixes from shadow-indexing-merged:
- 6bb1a5b fix use-after-free
- 591b01e get rid of remote_partition::reader_state
- 7a71158 correctly update delta
- 61ad458 reproduce and fix infinite cycle
- ef7a8a9 fix updating reader start offsets
- dfe705e remove _done
- 21a238f fix hydration in the remote segment
- 171ab0c remove timeout from work-item queue
- f916b8c prevent remote-segment from being evicted

Co-authored-by: Alexey Zatelepin <ztlpn@vectorized.io>