kafka/client: Fix concurrency bug with retry fn
- This was leading to all sorts of confusing to debug issues. The
behavior that was observed was invocations of `func` and `errFunc`
interleaving instead of occuring synchronously with respect to
eachother.

- The fact that eptr was being set meant that `errFunc` will be invoked,
however the erraneous reset to another local var that would never be
used meant that it was no longer to be invoked within the chain of calls
with the futures returned from the method.