rpc/transport: fixed error parsing response with error status
Our RPC implementation support propagating server errors to the client.
For this we use a `meta` field in RPC response header. The status that
is included into `header::meta` field indicates whether the request is
successful or its execution failed on the server side. When status
indicates an error response body is empty.

Fixed an error caused by the fact that we tried to parse response body
before we checked the status flag. This caused parsing error and broken
promise exception on the client side.

Signed-off-by: Michal Maslanka <michal@vectorized.io>