archival: Use upload_controller to control upload speed
Initialize upload_controller in the application.cc. Add configuration
parameters with reasonable defaults for it.

The proportional coefficient is set to 5 so we will stop throttling if
the upload backlog is at 20% of available disk space.
Integral and derivative coefficients are set to 0. The check is
performed every two minutes.