r/consensus: fixed validating configuration change in leader transfer
In redpanda raft implementation supports relaxed consistency. With
relaxed consistency we not always update `_commit_index` even tho
entries were replicated to all the nodes. When performing leadership
transfer we need to take that into account. When configuration is
successfully replicated to majority of nodes and it is not in joint
consensus state we are safe to proceed with the transfer.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 42d3d54f9a2d4c5e8e232d95cc6480c5fee422e6)