coproc: Reconciliation backend
- The reconciliation backend will mimmic the
cluster::controller_backend.

- Its purpose is to handle all non_replicable* op type events that are
emitted from the topics table.

- The rational for this is that it is not possible for coproc to
synchronize around the handling of these op_types if they are handled by
code in cluster. It can only know that the event was emitted, not when
it was finished.

- With this solution the complexity of the existing
cluster::controller_backend is reduced and more coproc code can stay in
the coproc namespace

- The process_update method is empty as this commit is just about
setting up the scaffolding that will handle events. Commits that
acutally handle the events are in subsequent commits within this patch
series