kafka: make list_offset linearizable
When we call consensus::is_leader the call uses cached state so
there is a possibility of returning stale data: the node still
thinks it's a leader while in fact it was isolated and the rest
of the cluster elected a new leader.

Switching is_leader with linearizable_barrier to make sure that
the current node is/was a leader when it started executing the
request.

Checked that linearizability was violated with chaos testing and
then checked that the fix fixed it.

Chaos tests: https://github.com/vectorizedio/chaos

Used writing_list_offsets/isolate_leader.json for validation

(cherry picked from commit 3aa33d37d69067f1690ddf979f2a61dd1ea346f2)